+++ narps_open/pipelines/team_X19V.py
@@ -1,8 +1,13 @@
+#!/usr/bin/python
+# coding: utf-8
+"""Classes and functions for the pipeline of team X19V."""
+
 import os
 import shutil
-from os.path import join as opj
+from os.path import join
+from pathlib import Path
 
-import nibabel as nib
+import nibabel as nb
 import numpy as np
 import pandas as pd
 from nipype import MapNode, Node, Workflow
@@ -24,942 +29,986 @@ from nipype.interfaces.fsl import (
 from nipype.interfaces.io import DataSink, SelectFiles
 from nipype.interfaces.utility import Function, IdentityInterface
 
+from narps_open.pipelines import Pipeline
+
 # Event-Related design, 4 second events, with parametric modulation based on amount gained or lost on each trial.
 # One 'gain' regressor, and one 'loss' regressor.
 # No RT modelling
 
 
-def get_session_infos(event_file):
-    """Create Bunchs for specifyModel.
+class PipelineTeamX19V(Pipeline):
+    """A class that defines the pipeline of team X19V."""
 
-    Parameters :
-    - event_file : str, file corresponding to the run and the subject to analyze
+    def __init__(self, bids_dir: str | Path, subject_list: str | list[str]):
+        """Create a pipeline object."""
+        super().__init__()
+        self.fwhm = 5.0
+        self.team_id = "X19V"
+        self.contrast_list = ["0001", "0002", "0003"]
 
-    Returns :
-    - subject_info : list of Bunch for 1st level analysis.
-    """
-    cond_names = ["trial", "gain", "loss"]
+        self.directories.dataset_dir = str(bids_dir)
 
-    onset = {}
-    duration = {}
-    amplitude = {}
+        if isinstance(subject_list, str):
+            subject_list = [subject_list]
+        self.subject_list = subject_list
 
-    for c in cond_names:  # For each condition.
-        onset.update({c: []})  # creates dictionary items with empty lists
-        duration.update({c: []})
-        amplitude.update({c: []})
+    def get_preprocessing():
+        """Create preprocessing workflow.
 
-    with open(event_file, "rt") as f:
-        next(f)  # skip the header
+        Unused by this team.
+        """
+        ...
 
-        for line in f:
-            info = line.strip().split()
-            # Creates list with onsets, duration and loss/gain for amplitude (FSL)
-            for c in cond_names:
-                onset[c].append(float(info[0]))
-                duration[c].append(float(4))
-                if c == "gain":
-                    amplitude[c].append(float(info[2]))
-                elif c == "loss":
-                    amplitude[c].append(float(info[3]))
-                elif c == "trial":
-                    amplitude[c].append(float(1))
+    def get_subject_infos(self, event_file: str) -> list[type[Bunch]]:
+        """Create Bunchs for specifyModel.
 
-    amplitude["gain"] = amplitude["gain"] - np.mean(amplitude["gain"])
-    amplitude["loss"] = amplitude["loss"] - np.mean(amplitude["loss"])
+        Parameters :
+        - event_file : str, file corresponding to the run and the subject to analyze
 
-    subject_info = []
+        Returns :
+        - subject_info : list of Bunch for 1st level analysis.
+        """
+        cond_names = ["trial", "gain", "loss"]
 
-    subject_info.append(
-        Bunch(
-            conditions=cond_names,
-            onsets=[onset[k] for k in cond_names],
-            durations=[duration[k] for k in cond_names],
-            amplitudes=[amplitude[k] for k in cond_names],
-            regressor_names=None,
-            regressors=None,
+        onset = {}
+        duration = {}
+        amplitude = {}
+
+        for c in cond_names:  # For each condition.
+            onset.update({c: []})  # creates dictionary items with empty lists
+            duration.update({c: []})
+            amplitude.update({c: []})
+
+        with open(event_file, "rt") as f:
+            next(f)  # skip the header
+
+            for line in f:
+                info = line.strip().split()
+                # Creates list with onsets, duration and loss/gain for amplitude (FSL)
+                for c in cond_names:
+                    onset[c].append(float(info[0]))
+                    duration[c].append(float(4))
+                    if c == "gain":
+                        amplitude[c].append(float(info[2]))
+                    elif c == "loss":
+                        amplitude[c].append(float(info[3]))
+                    elif c == "trial":
+                        amplitude[c].append(float(1))
+
+        amplitude["gain"] = amplitude["gain"] - np.mean(amplitude["gain"])
+        amplitude["loss"] = amplitude["loss"] - np.mean(amplitude["loss"])
+
+        subject_info = []
+
+        subject_info.append(
+            Bunch(
+                conditions=cond_names,
+                onsets=[onset[k] for k in cond_names],
+                durations=[duration[k] for k in cond_names],
+                amplitudes=[amplitude[k] for k in cond_names],
+                regressor_names=None,
+                regressors=None,
+            )
         )
-    )
 
-    return subject_info
+        return subject_info
 
+    def get_parameters_file(
+        self,
+        file: str | Path,
+        subject_id: str,
+        run_id: str,
+    ):
+        """Create new tsv files with only desired parameters per subject per run.
 
-def get_parameters_file(file, subject_id, run_id, result_dir, working_dir):
-    """Create new tsv files with only desired parameters per subject per run.
-
-    Parameters :
-    - filepaths : paths to subject parameters file (i.e. one per run)
-    - subject_id : subject for whom the 1st level analysis is made
-    - run_id: run for which the 1st level analysis is made
+        Parameters :
+        - filepaths : paths to subject parameters file (i.e. one per run)
+        - subject_id : subject for whom the 1st level analysis is made
+        - run_id: run for which the 1st level analysis is made
         - result_dir: str, directory where results will be stored
         - working_dir: str, name of the sub-directory for intermediate results
 
-    Return :
-    - parameters_file : paths to new files containing only desired parameters.
-    """
-    parameters_file = []
+        Return :
+        - parameters_file : paths to new files containing only desired parameters.
+        """
+        df = pd.read_csv(file, sep="\t", header=0)
+        temp_list = np.array(
+            [df["X"], df["Y"], df["Z"], df["RotX"], df["RotY"], df["RotZ"]]
+        )  # Parameters we want to use for the model
+        retained_parameters = pd.DataFrame(np.transpose(temp_list))
 
-    df = pd.read_csv(file, sep="\t", header=0)
-    temp_list = np.array(
-        [df["X"], df["Y"], df["Z"], df["RotX"], df["RotY"], df["RotZ"]]
-    )  # Parameters we want to use for the model
-    retained_parameters = pd.DataFrame(np.transpose(temp_list))
-    new_path = opj(
-        result_dir,
-        working_dir,
-        "parameters_file",
-        f"parameters_file_sub-{subject_id}_run{run_id}.tsv",
-    )
-    if not os.path.isdir(opj(result_dir, working_dir, "parameters_file")):
-        os.mkdir(opj(result_dir, working_dir, "parameters_file"))
-    writer = open(new_path, "w")
-    writer.write(
-        retained_parameters.to_csv(sep="\t", index=False, header=False, na_rep="0.0")
-    )
-    writer.close()
+        paramater_files_dir = Path(
+            self.directories.results_dir
+            / self.directories.working_dir
+            / "parameters_file"
+        )
+        paramater_files_dir.mkdir(parents=True, exist_ok=True)
 
-    parameters_file.append(new_path)
+        new_path = (
+            paramater_files_dir / f"parameters_file_sub-{subject_id}_run{run_id}.tsv"
+        )
+        retained_parameters.to_csv(
+            new_path, sep="\t", index=False, header=False, na_rep="0.0"
+        )
 
-    return parameters_file
+        parameters_file = [new_path]
+        return parameters_file
 
+    # [INFO] Linear contrast effects: 'Gain' vs. baseline, 'Loss' vs. baseline.
+    def get_contrasts(self) -> list[tuple]:
+        """Create the list of tuples that represents contrasts.
 
-# Linear contrast effects: 'Gain' vs. baseline, 'Loss' vs. baseline.
-def get_contrasts():
-    """
-    Create the list of tuples that represents contrasts.
-    Each contrast is in the form :
-    (Name,Stat,[list of condition names],[weights on those conditions])
+        Each contrast is in the form :
+        (Name, Stat, [list of condition names], [weights on those conditions])
 
-    Returns:
-        - contrasts: list of tuples, list of contrasts to analyze
-    """
-    # list of condition names
-    conditions = ["trial", "gain", "loss"]
+        Parameters:
+            - subject_id: str, ID of the subject
 
-    # create contrasts
-    gain = ("gain", "T", conditions, [0, 1, 0])
+        Returns:
+            - contrasts: list of tuples, list of contrasts to analyze
+        """
+        # list of condition names
+        conditions = ["trial", "gain", "loss"]
 
-    loss = ("loss", "T", conditions, [0, 0, 1])
+        # create contrasts
+        gain = ("gain", "T", conditions, [0, 1, 0])
 
-    gain_sup = ("gain_sup_loss", "T", conditions, [0, 1, -1])
+        loss = ("loss", "T", conditions, [0, 0, 1])
 
-    loss_sup = ("loss_sup_gain", "T", conditions, [0, -1, 1])
+        gain_sup = ("gain_sup_loss", "T", conditions, [0, 1, -1])
 
-    # contrast list
-    contrasts = [gain, loss, gain_sup, loss_sup]
+        loss_sup = ("loss_sup_gain", "T", conditions, [0, -1, 1])
 
-    return contrasts
+        # contrast list
+        contrasts = [gain, loss, gain_sup, loss_sup]
 
+        return contrasts
 
-def rm_smoothed_files(subject_id, run_id, result_dir, working_dir):
-    smooth_dir = opj(
-        result_dir,
-        working_dir,
-        "l1_analysis",
-        f"_run_id_{run_id}_subject_id_{subject_id}",
-        "smooth",
-    )
+    def get_run_level_analysis(self):
+        """Return the first level analysis workflow.
 
-    try:
-        shutil.rmtree(smooth_dir)
-    except OSError as e:
-        print(e)
-    else:
-        print("The directory is deleted successfully")
+        Parameters:
+                - result_dir: str, directory where results will be stored
+                - working_dir: str, name of the sub-directory for intermediate results
 
-
-def get_l1_analysis(
-    subject_list, run_list, TR, fwhm, exp_dir, output_dir, working_dir, result_dir
-):
-    """
-    Returns the first level analysis workflow.
-
-    Parameters:
-            - exp_dir: str, directory where raw data are stored
-            - result_dir: str, directory where results will be stored
-            - working_dir: str, name of the sub-directory for intermediate results
-            - output_dir: str, name of the sub-directory for final results
-            - subject_list: list of str, list of subject for which you want to do the analysis
-            - run_list: list of str, list of runs for which you want to do the analysis
-            - fwhm: float, fwhm for smoothing step
-            - TR: float, time repetition used during acquisition
-
-    Returns:
-            - l1_analysis : Nipype WorkFlow
-    """
-    # Infosource Node - To iterate on subject and runs
-    infosource = Node(
-        IdentityInterface(fields=["subject_id", "run_id"]), name="infosource"
-    )
-    infosource.iterables = [("subject_id", subject_list), ("run_id", run_list)]
-
-    # Templates to select files node
-    func_file = opj(
-        "derivatives",
-        "fmriprep",
-        "sub-{subject_id}",
-        "func",
-        "sub-{subject_id}_task-MGT_run-{run_id}_bold_space-MNI152NLin2009cAsym_preproc.nii.gz",
-    )
-
-    event_file = opj(
-        "sub-{subject_id}", "func", "sub-{subject_id}_task-MGT_run-{run_id}_events.tsv"
-    )
-
-    param_file = opj(
-        "derivatives",
-        "fmriprep",
-        "sub-{subject_id}",
-        "func",
-        "sub-{subject_id}_task-MGT_run-{run_id}_bold_confounds.tsv",
-    )
-
-    template = {"func": func_file, "event": event_file, "param": param_file}
-
-    # SelectFiles node - to select necessary files
-    selectfiles = Node(
-        SelectFiles(template, base_directory=exp_dir), name="selectfiles"
-    )
-
-    # DataSink Node - store the wanted results in the wanted repository
-    datasink = Node(
-        DataSink(base_directory=result_dir, container=output_dir), name="datasink"
-    )
-
-    # Skullstripping
-    skullstrip = Node(BET(frac=0.1, functional=True, mask=True), name="skullstrip")
-
-    # Smoothing
-    smooth = Node(IsotropicSmooth(fwhm=5), name="smooth")
-
-    # Node contrasts to get contrasts
-    contrasts = Node(
-        Function(
-            function=get_contrasts,
-            input_names=["subject_id"],
-            output_names=["contrasts"],
-        ),
-        name="contrasts",
-    )
-
-    # Get Subject Info - get subject specific condition information
-    get_subject_infos = Node(
-        Function(
-            input_names=["event_file"],
-            output_names=["subject_info"],
-            function=get_session_infos,
-        ),
-        name="get_subject_infos",
-    )
-
-    specify_model = Node(
-        SpecifyModel(
-            high_pass_filter_cutoff=60, input_units="secs", time_repetition=TR
-        ),
-        name="specify_model",
-    )
-
-    parameters = Node(
-        Function(
-            function=get_parameters_file,
-            input_names=["file", "subject_id", "run_id", "result_dir", "working_dir"],
-            output_names=["parameters_file"],
-        ),
-        name="parameters",
-    )
-
-    parameters.inputs.result_dir = result_dir
-    parameters.inputs.working_dir = working_dir
-
-    # First temporal derivatives of the two regressors were also used,
-    # along with temporal filtering (60 s) of all the independent variable time-series.
-    # No motion parameter regressors used.
-
-    l1_design = Node(
-        Level1Design(
-            bases={"dgamma": {"derivs": False}},
-            interscan_interval=TR,
-            model_serial_correlations=True,
-        ),
-        name="l1_design",
-    )
-
-    model_generation = Node(FEATModel(), name="model_generation")
-
-    model_estimate = Node(FILMGLS(), name="model_estimate")
-
-    remove_smooth = Node(
-        Function(
-            input_names=["subject_id", "run_id", "files", "result_dir", "working_dir"],
-            function=rm_smoothed_files,
-        ),
-        name="remove_smooth",
-    )
-
-    remove_smooth.inputs.result_dir = result_dir
-    remove_smooth.inputs.working_dir = working_dir
-
-    # Create l1 analysis workflow and connect its nodes
-    l1_analysis = Workflow(base_dir=opj(result_dir, working_dir), name="l1_analysis")
-
-    l1_analysis.connect(
-        [
-            (
-                infosource,
-                selectfiles,
-                [("subject_id", "subject_id"), ("run_id", "run_id")],
-            ),
-            (selectfiles, get_subject_infos, [("event", "event_file")]),
-            (selectfiles, parameters, [("param", "file")]),
-            (infosource, contrasts, [("subject_id", "subject_id")]),
-            (
-                infosource,
-                parameters,
-                [("subject_id", "subject_id"), ("run_id", "run_id")],
-            ),
-            (selectfiles, skullstrip, [("func", "in_file")]),
-            (skullstrip, smooth, [("out_file", "in_file")]),
-            (
-                parameters,
-                specify_model,
-                [("parameters_file", "realignment_parameters")],
-            ),
-            (smooth, specify_model, [("out_file", "functional_runs")]),
-            (get_subject_infos, specify_model, [("subject_info", "subject_info")]),
-            (contrasts, l1_design, [("contrasts", "contrasts")]),
-            (specify_model, l1_design, [("session_info", "session_info")]),
-            (
-                l1_design,
-                model_generation,
-                [("ev_files", "ev_files"), ("fsf_files", "fsf_file")],
-            ),
-            (smooth, model_estimate, [("out_file", "in_file")]),
-            (
-                model_generation,
-                model_estimate,
-                [("design_file", "design_file")],
-            ),
-            (
-                infosource,
-                remove_smooth,
-                [("subject_id", "subject_id"), ("run_id", "run_id")],
-            ),
-            (model_estimate, remove_smooth, [("results_dir", "files")]),
-            (model_estimate, datasink, [("results_dir", "l1_analysis.@results")]),
-            (
-                model_generation,
-                datasink,
-                [
-                    ("design_file", "l1_analysis.@design_file"),
-                    ("design_image", "l1_analysis.@design_img"),
-                ],
-            ),
-            (skullstrip, datasink, [("mask_file", "l1_analysis.@skullstriped")]),
+        Returns:
+                - l1_analysis : Nipype WorkFlow
+        """
+        # Infosource Node - To iterate on subject and runs
+        infosource = Node(
+            IdentityInterface(fields=["subject_id", "run_id"]), name="infosource"
+        )
+        infosource.iterables = [
+            ("subject_id", self.subject_list),
+            ("run_id", self.run_list),
         ]
-    )
 
-    return l1_analysis
+        # Templates to select files node
+        func_file = join(
+            "derivatives",
+            "fmriprep",
+            "sub-{subject_id}",
+            "func",
+            "sub-{subject_id}_task-MGT_run-{run_id}_bold_space-MNI152NLin2009cAsym_preproc.nii.gz",
+        )
+        event_file = join(
+            "sub-{subject_id}",
+            "func",
+            "sub-{subject_id}_task-MGT_run-{run_id}_events.tsv",
+        )
+        param_file = join(
+            "derivatives",
+            "fmriprep",
+            "sub-{subject_id}",
+            "func",
+            "sub-{subject_id}_task-MGT_run-{run_id}_bold_confounds.tsv",
+        )
+        template = {"func": func_file, "event": event_file, "param": param_file}
 
+        # SelectFiles node - to select necessary files
+        selectfiles = Node(
+            SelectFiles(template, base_directory=str(self.directories.dataset_dir)),
+            name="selectfiles",
+        )
 
-def get_l2_analysis(
-    subject_list, contrast_list, run_list, exp_dir, output_dir, working_dir, result_dir
-):
-    """
-    Returns the 2nd level of analysis workflow.
-
-    Parameters:
-            - exp_dir: str, directory where raw data are stored
-            - result_dir: str, directory where results will be stored
-            - working_dir: str, name of the sub-directory for intermediate results
-            - output_dir: str, name of the sub-directory for final results
-            - subject_list: list of str, list of subject for which you want to do the preprocessing
-            - contrast_list: list of str, list of contrasts to analyze
-            - run_list: list of str, list of runs for which you want to do the analysis
-
-
-    Returns:
-    - l2_analysis: Nipype WorkFlow
-    """
-    # Infosource Node - To iterate on subject and runs
-    infosource_2ndlevel = Node(
-        IdentityInterface(fields=["subject_id", "contrast_id"]),
-        name="infosource_2ndlevel",
-    )
-    infosource_2ndlevel.iterables = [
-        ("subject_id", subject_list),
-        ("contrast_id", contrast_list),
-    ]
-
-    # Templates to select files node
-    copes_file = opj(
-        output_dir,
-        "l1_analysis",
-        "_run_id_*_subject_id_{subject_id}",
-        "results",
-        "cope{contrast_id}.nii.gz",
-    )
-
-    varcopes_file = opj(
-        output_dir,
-        "l1_analysis",
-        "_run_id_*_subject_id_{subject_id}",
-        "results",
-        "varcope{contrast_id}.nii.gz",
-    )
-
-    mask_file = opj(
-        output_dir,
-        "l1_analysis",
-        "_run_id_*_subject_id_{subject_id}",
-        "sub-{subject_id}_task-MGT_run-01_bold_space-MNI152NLin2009cAsym_preproc_brain_mask.nii.gz",
-    )
-
-    template = {"cope": copes_file, "varcope": varcopes_file, "mask": mask_file}
-
-    # SelectFiles node - to select necessary files
-    selectfiles_2ndlevel = Node(
-        SelectFiles(template, base_directory=result_dir), name="selectfiles_2ndlevel"
-    )
-
-    datasink_2ndlevel = Node(
-        DataSink(base_directory=result_dir, container=output_dir),
-        name="datasink_2ndlevel",
-    )
-
-    # Generate design matrix
-    specify_model_2ndlevel = Node(
-        L2Model(num_copes=len(run_list)), name="l2model_2ndlevel"
-    )
-
-    # Merge copes and varcopes files for each subject
-    merge_copes_2ndlevel = Node(Merge(dimension="t"), name="merge_copes_2ndlevel")
-
-    merge_varcopes_2ndlevel = Node(Merge(dimension="t"), name="merge_varcopes_2ndlevel")
-
-    # Second level (single-subject, mean of all four scans) analyses: Fixed effects analysis.
-    flame = Node(FLAMEO(run_mode="flame1"), name="flameo")
-
-    l2_analysis = Workflow(base_dir=opj(result_dir, working_dir), name="l2_analysis")
-
-    l2_analysis.connect(
-        [
-            (
-                infosource_2ndlevel,
-                selectfiles_2ndlevel,
-                [("subject_id", "subject_id"), ("contrast_id", "contrast_id")],
+        # DataSink Node - store the wanted results in the wanted repository
+        datasink = Node(
+            DataSink(
+                base_directory=str(self.directories.results_dir),
+                container=self.directories.output_dir,
             ),
-            (selectfiles_2ndlevel, merge_copes_2ndlevel, [("cope", "in_files")]),
-            (selectfiles_2ndlevel, merge_varcopes_2ndlevel, [("varcope", "in_files")]),
-            (selectfiles_2ndlevel, flame, [("mask", "mask_file")]),
-            (merge_copes_2ndlevel, flame, [("merged_file", "cope_file")]),
-            (merge_varcopes_2ndlevel, flame, [("merged_file", "var_cope_file")]),
-            (
-                specify_model_2ndlevel,
-                flame,
-                [
-                    ("design_mat", "design_file"),
-                    ("design_con", "t_con_file"),
-                    ("design_grp", "cov_split_file"),
+            name="datasink",
+        )
+
+        # Skullstripping
+        skullstrip = Node(BET(frac=0.1, functional=True, mask=True), name="skullstrip")
+
+        # Smoothing
+        smooth = Node(IsotropicSmooth(fwhm=self.fwhm), name="smooth")
+
+        # Node contrasts to get contrasts
+        contrasts = Node(
+            Function(
+                function=self.get_contrasts,
+                input_names=["subject_id"],
+                output_names=["contrasts"],
+            ),
+            name="contrasts",
+        )
+
+        # Get Subject Info - get subject specific condition information
+        get_subject_infos = Node(
+            Function(
+                input_names=["event_file"],
+                output_names=["subject_info"],
+                function=self.get_subject_infos,
+            ),
+            name="get_subject_infos",
+        )
+
+        specify_model = Node(
+            SpecifyModel(
+                high_pass_filter_cutoff=60, input_units="secs", time_repetition=self.tr
+            ),
+            name="specify_model",
+        )
+
+        parameters = Node(
+            Function(
+                function=self.get_parameters_file,
+                input_names=[
+                    "file",
+                    "subject_id",
+                    "run_id",
+                    "result_dir",
+                    "working_dir",
                 ],
+                output_names=["parameters_file"],
             ),
-            (
-                flame,
-                datasink_2ndlevel,
-                [
-                    ("zstats", "l2_analysis.@stats"),
-                    ("tstats", "l2_analysis.@tstats"),
-                    ("copes", "l2_analysis.@copes"),
-                    ("var_copes", "l2_analysis.@varcopes"),
+            name="parameters",
+        )
+
+        parameters.inputs.result_dir = self.directories.results_dir
+        parameters.inputs.working_dir = self.directories.working_dir
+
+        # First temporal derivatives of the two regressors were also used,
+        # along with temporal filtering (60 s) of all the independent variable time-series.
+        # No motion parameter regressors used.
+
+        l1_design = Node(
+            Level1Design(
+                bases={"dgamma": {"derivs": False}},
+                interscan_interval=self.tr,
+                model_serial_correlations=True,
+            ),
+            name="l1_design",
+        )
+
+        model_generation = Node(FEATModel(), name="model_generation")
+
+        model_estimate = Node(FILMGLS(), name="model_estimate")
+
+        remove_smooth = Node(
+            Function(
+                input_names=[
+                    "subject_id",
+                    "run_id",
+                    "files",
+                    "result_dir",
+                    "working_dir",
                 ],
+                function=self.rm_smoothed_files,
             ),
+            name="remove_smooth",
+        )
+
+        remove_smooth.inputs.result_dir = self.directories.results_dir
+        remove_smooth.inputs.working_dir = self.directories.working_dir
+
+        # Create l1 analysis workflow and connect its nodes
+        l1_analysis = Workflow(
+            base_dir=join(self.directories.results_dir, self.directories.working_dir),
+            name="l1_analysis",
+        )
+
+        l1_analysis.connect(
+            [
+                (
+                    infosource,
+                    selectfiles,
+                    [("subject_id", "subject_id"), ("run_id", "run_id")],
+                ),
+                (selectfiles, get_subject_infos, [("event", "event_file")]),
+                (selectfiles, parameters, [("param", "file")]),
+                (infosource, contrasts, [("subject_id", "subject_id")]),
+                (
+                    infosource,
+                    parameters,
+                    [("subject_id", "subject_id"), ("run_id", "run_id")],
+                ),
+                (selectfiles, skullstrip, [("func", "in_file")]),
+                (skullstrip, smooth, [("out_file", "in_file")]),
+                (
+                    parameters,
+                    specify_model,
+                    [("parameters_file", "realignment_parameters")],
+                ),
+                (smooth, specify_model, [("out_file", "functional_runs")]),
+                (get_subject_infos, specify_model, [("subject_info", "subject_info")]),
+                (contrasts, l1_design, [("contrasts", "contrasts")]),
+                (specify_model, l1_design, [("session_info", "session_info")]),
+                (
+                    l1_design,
+                    model_generation,
+                    [("ev_files", "ev_files"), ("fsf_files", "fsf_file")],
+                ),
+                (smooth, model_estimate, [("out_file", "in_file")]),
+                (
+                    model_generation,
+                    model_estimate,
+                    [("design_file", "design_file")],
+                ),
+                (
+                    infosource,
+                    remove_smooth,
+                    [("subject_id", "subject_id"), ("run_id", "run_id")],
+                ),
+                (model_estimate, remove_smooth, [("results_dir", "files")]),
+                (model_estimate, datasink, [("results_dir", "l1_analysis.@results")]),
+                (
+                    model_generation,
+                    datasink,
+                    [
+                        ("design_file", "l1_analysis.@design_file"),
+                        ("design_image", "l1_analysis.@design_img"),
+                    ],
+                ),
+                (skullstrip, datasink, [("mask_file", "l1_analysis.@skullstriped")]),
+            ]
+        )
+
+        return l1_analysis
+
+    def get_subject_level_analysis(self):
+        """Return the 2nd level of analysis workflow.
+
+        Returns:
+        - l2_analysis: Nipype WorkFlow
+        """
+        # Infosource Node - To iterate on subject and runs
+        infosource_2ndlevel = Node(
+            IdentityInterface(fields=["subject_id", "contrast_id"]),
+            name="infosource_2ndlevel",
+        )
+        infosource_2ndlevel.iterables = [
+            ("subject_id", self.subject_list),
+            ("contrast_id", self.contrast_list),
         ]
-    )
 
-    return l2_analysis
+        # Templates to select files node
+        copes_file = join(
+            self.directories.output_dir,
+            "l1_analysis",
+            "_run_id_*_subject_id_{subject_id}",
+            "results",
+            "cope{contrast_id}.nii.gz",
+        )
+        varcopes_file = join(
+            self.directories.output_dir,
+            "l1_analysis",
+            "_run_id_*_subject_id_{subject_id}",
+            "results",
+            "varcope{contrast_id}.nii.gz",
+        )
+        mask_file = join(
+            self.directories.output_dir,
+            "l1_analysis",
+            "_run_id_*_subject_id_{subject_id}",
+            "sub-{subject_id}_task-MGT_run-01_bold_space-MNI152NLin2009cAsym_preproc_brain_mask.nii.gz",
+        )
+        template = {"cope": copes_file, "varcope": varcopes_file, "mask": mask_file}
 
+        # SelectFiles node - to select necessary files
+        selectfiles_2ndlevel = Node(
+            SelectFiles(template, base_directory=str(self.directories.results_dir)),
+            name="selectfiles_2ndlevel",
+        )
 
-def get_subgroups_contrasts(copes, varcopes, subject_ids, participants_file):
-    """
-    Parameters :
-    - copes: original file list selected by selectfiles node
-    - varcopes: original file list selected by selectfiles node
-    - subject_ids: list of subject IDs that are analyzed
-    - participants_file: str, file containing participants characteristics
+        datasink_2ndlevel = Node(
+            DataSink(
+                base_directory=str(self.directories.results_dir),
+                container=self.directories.output_dir,
+            ),
+            name="datasink_2ndlevel",
+        )
 
-    This function return the file list containing only the files belonging to subject in the wanted group.
-    """
-    equalRange_id = []
-    equalIndifference_id = []
+        # Generate design matrix
+        specify_model_2ndlevel = Node(
+            L2Model(num_copes=len(self.run_list)), name="l2model_2ndlevel"
+        )
 
-    subject_list = ["sub-" + str(i) for i in subject_ids]
+        # Merge copes and varcopes files for each subject
+        merge_copes_2ndlevel = Node(Merge(dimension="t"), name="merge_copes_2ndlevel")
+        merge_varcopes_2ndlevel = Node(
+            Merge(dimension="t"), name="merge_varcopes_2ndlevel"
+        )
 
-    with open(participants_file, "rt") as f:
-        next(f)  # skip the header
+        # Second level (single-subject, mean of all four scans) analyses: Fixed effects analysis.
+        flame = Node(FLAMEO(run_mode="flame1"), name="flameo")
 
-        for line in f:
-            info = line.strip().split()
+        l2_analysis = Workflow(
+            base_dir=join(self.directories.results_dir, self.directories.working_dir),
+            name="l2_analysis",
+        )
 
-            if info[0] in subject_list and info[1] == "equalIndifference":
-                equalIndifference_id.append(info[0][-3:])
-            elif info[0] in subject_list and info[1] == "equalRange":
-                equalRange_id.append(info[0][-3:])
+        l2_analysis.connect(
+            [
+                (
+                    infosource_2ndlevel,
+                    selectfiles_2ndlevel,
+                    [("subject_id", "subject_id"), ("contrast_id", "contrast_id")],
+                ),
+                (selectfiles_2ndlevel, merge_copes_2ndlevel, [("cope", "in_files")]),
+                (
+                    selectfiles_2ndlevel,
+                    merge_varcopes_2ndlevel,
+                    [("varcope", "in_files")],
+                ),
+                (selectfiles_2ndlevel, flame, [("mask", "mask_file")]),
+                (merge_copes_2ndlevel, flame, [("merged_file", "cope_file")]),
+                (merge_varcopes_2ndlevel, flame, [("merged_file", "var_cope_file")]),
+                (
+                    specify_model_2ndlevel,
+                    flame,
+                    [
+                        ("design_mat", "design_file"),
+                        ("design_con", "t_con_file"),
+                        ("design_grp", "cov_split_file"),
+                    ],
+                ),
+                (
+                    flame,
+                    datasink_2ndlevel,
+                    [
+                        ("zstats", "l2_analysis.@stats"),
+                        ("tstats", "l2_analysis.@tstats"),
+                        ("copes", "l2_analysis.@copes"),
+                        ("var_copes", "l2_analysis.@varcopes"),
+                    ],
+                ),
+            ]
+        )
 
-    copes_equalIndifference = []
-    copes_equalRange = []
-    varcopes_equalIndifference = []
-    varcopes_equalRange = []
+        return l2_analysis
 
-    for file in copes:
-        sub_id = file.split("/")
-        print(sub_id)
-        if sub_id[-2][-3:] in equalIndifference_id:
-            copes_equalIndifference.append(file)
-        elif sub_id[-2][-3:] in equalRange_id:
-            copes_equalRange.append(file)
+    # [INFO] This function returns the list of ids and files of each group of participants
+    # to do analyses for both groups, and one between the two groups.
+    def get_subgroups_contrasts(
+        self,
+        copes: list[str],
+        varcopes: list[str],
+    ):
+        """Return the list of ids and files of each group of participants \
+           to do analyses for both groups, and one between the two groups.
 
-    for file in varcopes:
-        sub_id = file.split("/")
-        if sub_id[-2][-3:] in equalIndifference_id:
-            varcopes_equalIndifference.append(file)
-        elif sub_id[-2][-3:] in equalRange_id:
-            varcopes_equalRange.append(file)
+        Parameters :
+        - copes: original file list selected by selectfiles node
+        - varcopes: original file list selected by selectfiles node
 
-    print(len(equalRange_id))
-    print(len(equalIndifference_id))
-    print(len(copes_equalIndifference))
-    print(len(copes_equalRange))
+        This function return the file list containing
+        only the files belonging to subject in the wanted group.
+        """
+        equalRange_id = []
+        equalIndifference_id = []
 
-    copes_global = copes_equalIndifference + copes_equalRange
-    varcopes_global = varcopes_equalIndifference + varcopes_equalRange
+        subject_list = [f"sub-{str(i)}" for i in self.subject_list]
 
-    return (
-        copes_equalIndifference,
-        copes_equalRange,
-        copes_global,
-        varcopes_equalIndifference,
-        varcopes_equalRange,
-        varcopes_global,
-        equalIndifference_id,
-        equalRange_id,
-    )
+        participants_file = Path(self.directories.dataset_dir) / "participants.tsv"
 
+        with open(participants_file, "rt") as f:
+            next(f)  # skip the header
 
-def get_regs(equalRange_id, equalIndifference_id, method, subject_list):
-    """
-    Create dictionary of regressors for group analysis.
+            for line in f:
+                info = line.strip().split()
 
-    Parameters:
-            - equalRange_id: list of str, ids of subjects in equal range group
-            - equalIndifference_id: list of str, ids of subjects in equal indifference group
+                if info[0] in subject_list:
+                    subject_label = info[0][-3:]
+
+                    if info[1] == "equalIndifference":
+                        equalIndifference_id.append(subject_label)
+                    elif info[1] == "equalRange":
+                        equalRange_id.append(subject_label)
+
+        copes_equalIndifference = []
+        copes_equalRange = []
+        varcopes_equalIndifference = []
+        varcopes_equalRange = []
+
+        for file in copes:
+            sub_id = file.split("/")
+            subject_label = sub_id[-2][-3:]
+            if subject_label in equalIndifference_id:
+                copes_equalIndifference.append(file)
+            elif subject_label in equalRange_id:
+                copes_equalRange.append(file)
+
+        for file in varcopes:
+            sub_id = file.split("/")
+            subject_label = sub_id[-2][-3:]
+            if subject_label in equalIndifference_id:
+                varcopes_equalIndifference.append(file)
+            elif subject_label in equalRange_id:
+                varcopes_equalRange.append(file)
+
+        print(len(equalRange_id))
+        print(len(equalIndifference_id))
+        print(len(copes_equalIndifference))
+        print(len(copes_equalRange))
+
+        copes_global = copes_equalIndifference + copes_equalRange
+        varcopes_global = varcopes_equalIndifference + varcopes_equalRange
+
+        return (
+            copes_equalIndifference,
+            copes_equalRange,
+            copes_global,
+            varcopes_equalIndifference,
+            varcopes_equalRange,
+            varcopes_global,
+            equalIndifference_id,
+            equalRange_id,
+        )
+
+    # [INFO] This function creates the dictionary of regressors used in FSL Nipype pipelines
+    def get_regressors(
+        self,
+        equal_range_id: list,
+        equal_indifference_id: list,
+        method: str,
+    ) -> dict:
+        """Create dictionary of regressors for group analysis.
+
+        Parameters:
+            - equal_range_id: ids of subjects in equal range group
+            - equal_indifference_id: ids of subjects in equal indifference group
             - method: one of "equalRange", "equalIndifference" or "groupComp"
-            - subject_list: list of str, ids of subjects for which we do the analysis
 
-    Returns:
-            - regressors: dict, dictionary of regressors used to distinguish groups in FSL group analysis
-    """
-    if method == "equalRange":
-        regressors = dict(group_mean=[1 for i in range(len(equalRange_id))])
+        Returns:
+            - regressors: regressors used to distinguish groups in FSL group analysis
+        """
+        # For one sample t-test, creates a dictionary
+        # with a list of the size of the number of participants
+        if method == "equalRange":
+            regressors = dict(group_mean=[1 for _ in range(len(equal_range_id))])
+        elif method == "equalIndifference":
+            regressors = dict(group_mean=[1 for _ in range(len(equal_indifference_id))])
 
-    elif method == "equalIndifference":
-        regressors = dict(group_mean=[1 for i in range(len(equalIndifference_id))])
+        # For two sample t-test, creates 2 lists:
+        #  - one for equal range group,
+        #  - one for equal indifference group
+        # Each list contains n_sub values with 0 and 1 depending on the group of the participant
+        # For equalRange_reg list --> participants with a 1 are in the equal range group
+        elif method == "groupComp":
+            equalRange_reg = [
+                1 for _ in range(len(equal_range_id) + len(equal_indifference_id))
+            ]
+            equalIndifference_reg = [
+                0 for _ in range(len(equal_range_id) + len(equal_indifference_id))
+            ]
 
-    elif method == "groupComp":
-        equalRange_reg = [
-            1 for i in range(len(equalRange_id) + len(equalIndifference_id))
-        ]
-        equalIndifference_reg = [
-            0 for i in range(len(equalRange_id) + len(equalIndifference_id))
-        ]
+            for index, subject_id in enumerate(self.subject_list):
+                if subject_id in equal_indifference_id:
+                    equalIndifference_reg[index] = 1
+                    equalRange_reg[index] = 0
 
-        for i, sub_id in enumerate(subject_list):
-            if sub_id in equalIndifference_id:
-                index = i
-                equalIndifference_reg[index] = 1
-                equalRange_reg[index] = 0
+            regressors = dict(
+                equalRange=equalRange_reg, equalIndifference=equalIndifference_reg
+            )
 
-        regressors = dict(
-            equalRange=equalRange_reg, equalIndifference=equalIndifference_reg
+        return regressors
+
+    def rm_smoothed_files(self, subject_id: str, run_id: str):
+        """Remove directories containting smoothed files."""
+        smooth_dir = (
+            Path(self.directories.results_dir)
+            / self.directories.working_dir
+            / "l1_analysis"
+            / f"_run_id_{run_id}_subject_id_{subject_id}"
+            / "smooth"
         )
 
-    return regressors
+        try:
+            shutil.rmtree(smooth_dir)
+        except OSError as e:
+            print(e)
+        else:
+            print("The directory was deleted successfully")
+
+    def get_group_level_analysis(
+        self,
+        method,
+    ):
+        """Return the group level of analysis workflow.
+
+        Parameters:
+                - method: one of "equalRange", "equalIndifference" or "groupComp"
 
 
-def get_group_workflow(
-    subject_list,
-    n_sub,
-    contrast_list,
-    method,
-    exp_dir,
-    output_dir,
-    working_dir,
-    result_dir,
-    data_dir,
-):
-    """
-    Returns the group level of analysis workflow.
+        Returns:
+                - l3_analysis: Nipype WorkFlow
+        """
 
-    Parameters:
-            - exp_dir: str, directory where raw data are stored
-            - result_dir: str, directory where results will be stored
-            - working_dir: str, name of the sub-directory for intermediate results
-            - output_dir: str, name of the sub-directory for final results
-            - subject_list: list of str, list of subject for which you want to do the preprocessing
-            - contrast_list: list of str, list of contrasts to analyze
-            - method: one of "equalRange", "equalIndifference" or "groupComp"
-            - n_sub: int, number of subject to include
-
-    Returns:
-            - l2_analysis: Nipype WorkFlow
-    """
-    # Infosource Node - To iterate on subject and runs
-    infosource_3rdlevel = Node(
-        IdentityInterface(
-            fields=[
-                "contrast_id",
-                "exp_dir",
-                "result_dir",
-                "output_dir",
-                "working_dir",
-                "subject_list",
-                "method",
-            ],
-            exp_dir=exp_dir,
-            result_dir=result_dir,
-            output_dir=output_dir,
-            working_dir=working_dir,
-            subject_list=subject_list,
-            method=method,
-        ),
-        name="infosource_3rdlevel",
-    )
-    infosource_3rdlevel.iterables = [("contrast_id", contrast_list)]
-
-    # Templates to select files node
-    copes_file = opj(
-        output_dir,
-        "l2_analysis",
-        "_contrast_id_{contrast_id}_subject_id_*",
-        "cope1.nii.gz",
-    )
-
-    varcopes_file = opj(
-        output_dir,
-        "l2_analysis",
-        "_contrast_id_{contrast_id}_subject_id_*",
-        "varcope1.nii.gz",
-    )
-
-    participants_file = opj(exp_dir, "participants.tsv")
-
-    mask_file = opj(data_dir, "NARPS-X19V", "hypo1_unthresh.nii.gz")
-
-    template = {
-        "cope": copes_file,
-        "varcope": varcopes_file,
-        "participants": participants_file,
-        "mask": mask_file,
-    }
-
-    # SelectFiles node - to select necessary files
-    selectfiles_3rdlevel = Node(
-        SelectFiles(template, base_directory=result_dir), name="selectfiles_3rdlevel"
-    )
-
-    datasink_3rdlevel = Node(
-        DataSink(base_directory=result_dir, container=output_dir),
-        name="datasink_3rdlevel",
-    )
-
-    merge_copes_3rdlevel = Node(Merge(dimension="t"), name="merge_copes_3rdlevel")
-    merge_varcopes_3rdlevel = Node(Merge(dimension="t"), name="merge_varcopes_3rdlevel")
-
-    subgroups_contrasts = Node(
-        Function(
-            input_names=["copes", "varcopes", "subject_ids", "participants_file"],
-            output_names=[
-                "copes_equalIndifference",
-                "copes_equalRange",
-                "copes_global",
-                "varcopes_equalIndifference",
-                "varcopes_equalRange",
-                "varcopes_global",
-                "equalIndifference_id",
-                "equalRange_id",
-            ],
-            function=get_subgroups_contrasts,
-        ),
-        name="subgroups_contrasts",
-    )
-
-    specifymodel_3rdlevel = Node(MultipleRegressDesign(), name="specifymodel_3rdlevel")
-
-    flame_3rdlevel = Node(FLAMEO(run_mode="flame1"), name="flame_3rdlevel")
-
-    regs = Node(
-        Function(
-            input_names=[
-                "equalRange_id",
-                "equalIndifference_id",
-                "method",
-                "subject_list",
-            ],
-            output_names=["regressors"],
-            function=self.get_regs,
-        ),
-        name="regs",
-    )
-    regs.inputs.method = method
-    regs.inputs.subject_list = subject_list
-
-    smoothest = MapNode(SmoothEstimate(), name="smoothest", iterfield=["zstat_file"])
-
-    cluster = MapNode(
-        Cluster(
-            threshold=2.3, out_threshold_file=True, out_pval_file=True, pthreshold=0.05
-        ),
-        name="cluster",
-        iterfield=["in_file", "dlh", "volume", "cope_file"],
-        synchronize=True,
-    )
-
-    l3_analysis = Workflow(
-        base_dir=opj(result_dir, working_dir), name=f"l3_analysis_{method}_nsub_{n_sub}"
-    )
-
-    l3_analysis.connect(
-        [
-            (
-                infosource_3rdlevel,
-                selectfiles_3rdlevel,
-                [("contrast_id", "contrast_id")],
-            ),
-            (
-                infosource_3rdlevel,
-                subgroups_contrasts,
-                [("subject_list", "subject_ids")],
-            ),
-            (
-                selectfiles_3rdlevel,
-                subgroups_contrasts,
-                [
-                    ("cope", "copes"),
-                    ("varcope", "varcopes"),
-                    ("participants", "participants_file"),
+        # Infosource Node - To iterate on subject and runs
+        infosource_3rdlevel = Node(
+            IdentityInterface(
+                fields=[
+                    "contrast_id",
+                    "exp_dir",
+                    "result_dir",
+                    "output_dir",
+                    "working_dir",
+                    "subject_list",
+                    "method",
                 ],
+                exp_dir=self.directories.dataset_dir,
+                result_dir=self.directories.results_dir,
+                output_dir=self.directories.output_dir,
+                working_dir=self.directories.working_dir,
+                subject_list=self.subject_list,
+                method=method,
             ),
-            (selectfiles_3rdlevel, flame_3rdlevel, [("mask", "mask_file")]),
-            (selectfiles_3rdlevel, smoothest, [("mask", "mask_file")]),
-            (
-                subgroups_contrasts,
-                regs,
-                [
-                    ("equalRange_id", "equalRange_id"),
-                    ("equalIndifference_id", "equalIndifference_id"),
-                ],
-            ),
-            (regs, specifymodel_3rdlevel, [("regressors", "regressors")]),
-        ]
-    )
+            name="infosource_3rdlevel",
+        )
+        infosource_3rdlevel.iterables = [("contrast_id", self.contrast_list)]
 
-    if method == "equalIndifference":
-        specifymodel_3rdlevel.inputs.contrasts = [
-            ["group_mean", "T", ["group_mean"], [1]],
-            ["group_mean_neg", "T", ["group_mean"], [-1]],
-        ]
+        # Templates to select files node
+        copes_file = join(
+            self.directories.output_dir,
+            "l2_analysis",
+            "_contrast_id_{contrast_id}_subject_id_*",
+            "cope1.nii.gz",
+        )
+
+        varcopes_file = join(
+            self.directories.output_dir,
+            "l2_analysis",
+            "_contrast_id_{contrast_id}_subject_id_*",
+            "varcope1.nii.gz",
+        )
+
+        participants_file = join(self.directories.dataset_dir, "participants.tsv")
+
+        mask_file = join(
+            self.directories.dataset_dir, "..", "NARPS-X19V", "hypo1_unthresh.nii.gz"
+        )
+
+        template = {
+            "cope": copes_file,
+            "varcope": varcopes_file,
+            "participants": participants_file,
+            "mask": mask_file,
+        }
+
+        # SelectFiles node - to select necessary files
+        selectfiles_3rdlevel = Node(
+            SelectFiles(template, base_directory=str(self.directories.results_dir)),
+            name="selectfiles_3rdlevel",
+        )
+
+        datasink_3rdlevel = Node(
+            DataSink(
+                base_directory=str(self.directories.results_dir),
+                container=self.directories.output_dir,
+            ),
+            name="datasink_3rdlevel",
+        )
+
+        merge_copes_3rdlevel = Node(Merge(dimension="t"), name="merge_copes_3rdlevel")
+        merge_varcopes_3rdlevel = Node(
+            Merge(dimension="t"), name="merge_varcopes_3rdlevel"
+        )
+
+        subgroups_contrasts = Node(
+            Function(
+                input_names=["copes", "varcopes", "subject_ids", "participants_file"],
+                output_names=[
+                    "copes_equalIndifference",
+                    "copes_equalRange",
+                    "copes_global",
+                    "varcopes_equalIndifference",
+                    "varcopes_equalRange",
+                    "varcopes_global",
+                    "equalIndifference_id",
+                    "equalRange_id",
+                ],
+                function=self.get_subgroups_contrasts,
+            ),
+            name="subgroups_contrasts",
+        )
+
+        specifymodel_3rdlevel = Node(
+            MultipleRegressDesign(), name="specifymodel_3rdlevel"
+        )
+
+        flame_3rdlevel = Node(FLAMEO(run_mode="flame1"), name="flame_3rdlevel")
+
+        regs = Node(
+            Function(
+                input_names=[
+                    "equalRange_id",
+                    "equalIndifference_id",
+                    "method",
+                    "subject_list",
+                ],
+                output_names=["regressors"],
+                function=self.get_regressors,
+            ),
+            name="regs",
+        )
+        regs.inputs.method = method
+        regs.inputs.subject_list = self.subject_list
+
+        smoothest = MapNode(
+            SmoothEstimate(), name="smoothest", iterfield=["zstat_file"]
+        )
+
+        cluster = MapNode(
+            Cluster(
+                threshold=2.3,
+                out_threshold_file=True,
+                out_pval_file=True,
+                pthreshold=0.05,
+            ),
+            name="cluster",
+            iterfield=["in_file", "dlh", "volume", "cope_file"],
+            synchronize=True,
+        )
+
+        n_sub = len(self.subject_list)
+        l3_analysis = Workflow(
+            base_dir=join(self.directories.results_dir, self.directories.working_dir),
+            name=f"l3_analysis_{method}_nsub_{n_sub}",
+        )
 
         l3_analysis.connect(
             [
                 (
-                    subgroups_contrasts,
-                    merge_copes_3rdlevel,
-                    [("copes_equalIndifference", "in_files")],
+                    infosource_3rdlevel,
+                    selectfiles_3rdlevel,
+                    [("contrast_id", "contrast_id")],
                 ),
                 (
+                    infosource_3rdlevel,
                     subgroups_contrasts,
-                    merge_varcopes_3rdlevel,
-                    [("varcopes_equalIndifference", "in_files")],
+                    [("subject_list", "subject_ids")],
                 ),
+                (
+                    selectfiles_3rdlevel,
+                    subgroups_contrasts,
+                    [
+                        ("cope", "copes"),
+                        ("varcope", "varcopes"),
+                        ("participants", "participants_file"),
+                    ],
+                ),
+                (selectfiles_3rdlevel, flame_3rdlevel, [("mask", "mask_file")]),
+                (selectfiles_3rdlevel, smoothest, [("mask", "mask_file")]),
+                (
+                    subgroups_contrasts,
+                    regs,
+                    [
+                        ("equalRange_id", "equalRange_id"),
+                        ("equalIndifference_id", "equalIndifference_id"),
+                    ],
+                ),
+                (regs, specifymodel_3rdlevel, [("regressors", "regressors")]),
             ]
         )
-    elif method == "equalRange":
-        specifymodel_3rdlevel.inputs.contrasts = [
-            ["group_mean", "T", ["group_mean"], [1]],
-            ["group_mean_neg", "T", ["group_mean"], [-1]],
-        ]
+
+        if method == "equalIndifference" or method == "equalRange":
+            specifymodel_3rdlevel.inputs.contrasts = [
+                ["group_mean", "T", ["group_mean"], [1]],
+                ["group_mean_neg", "T", ["group_mean"], [-1]],
+            ]
+
+            if method == "equalIndifference":
+                l3_analysis.connect(
+                    [
+                        (
+                            subgroups_contrasts,
+                            merge_copes_3rdlevel,
+                            [("copes_equalIndifference", "in_files")],
+                        ),
+                        (
+                            subgroups_contrasts,
+                            merge_varcopes_3rdlevel,
+                            [("varcopes_equalIndifference", "in_files")],
+                        ),
+                    ]
+                )
+            elif method == "equalRange":
+                l3_analysis.connect(
+                    [
+                        (
+                            subgroups_contrasts,
+                            merge_copes_3rdlevel,
+                            [("copes_equalRange", "in_files")],
+                        ),
+                        (
+                            subgroups_contrasts,
+                            merge_varcopes_3rdlevel,
+                            [("varcopes_equalRange", "in_files")],
+                        ),
+                    ]
+                )
+
+        elif method == "groupComp":
+            specifymodel_3rdlevel.inputs.contrasts = [
+                ["equalRange_sup", "T", ["equalRange", "equalIndifference"], [1, -1]]
+            ]
+
+            l3_analysis.connect(
+                [
+                    (
+                        subgroups_contrasts,
+                        merge_copes_3rdlevel,
+                        [("copes_global", "in_files")],
+                    ),
+                    (
+                        subgroups_contrasts,
+                        merge_varcopes_3rdlevel,
+                        [("varcopes_global", "in_files")],
+                    ),
+                ]
+            )
 
         l3_analysis.connect(
             [
+                (merge_copes_3rdlevel, flame_3rdlevel, [("merged_file", "cope_file")]),
                 (
-                    subgroups_contrasts,
-                    merge_copes_3rdlevel,
-                    [("copes_equalRange", "in_files")],
+                    merge_varcopes_3rdlevel,
+                    flame_3rdlevel,
+                    [("merged_file", "var_cope_file")],
                 ),
                 (
-                    subgroups_contrasts,
-                    merge_varcopes_3rdlevel,
-                    [("varcopes_equalRange", "in_files")],
+                    specifymodel_3rdlevel,
+                    flame_3rdlevel,
+                    [
+                        ("design_mat", "design_file"),
+                        ("design_con", "t_con_file"),
+                        ("design_grp", "cov_split_file"),
+                    ],
+                ),
+                (
+                    flame_3rdlevel,
+                    cluster,
+                    [("zstats", "in_file"), ("copes", "cope_file")],
+                ),
+                (flame_3rdlevel, smoothest, [("zstats", "zstat_file")]),
+                (smoothest, cluster, [("dlh", "dlh"), ("volume", "volume")]),
+                (
+                    flame_3rdlevel,
+                    datasink_3rdlevel,
+                    [
+                        ("zstats", f"l3_analysis_{method}_nsub_{n_sub}.@zstats"),
+                        ("tstats", f"l3_analysis_{method}_nsub_{n_sub}.@tstats"),
+                    ],
+                ),
+                (
+                    cluster,
+                    datasink_3rdlevel,
+                    [
+                        (
+                            "threshold_file",
+                            f"l3_analysis_{method}_nsub_{n_sub}.@thresh",
+                        ),
+                        ("pval_file", f"l3_analysis_{method}_nsub_{n_sub}.@pval"),
+                    ],
                 ),
             ]
         )
 
-    elif method == "groupComp":
-        specifymodel_3rdlevel.inputs.contrasts = [
-            ["equalRange_sup", "T", ["equalRange", "equalIndifference"], [1, -1]]
+        return l3_analysis
+
+    def get_hypotheses_outputs():
+        """Get output for each hypothesis.
+
+        Not yet implemented.
+        """
+        ...
+
+    def reorganize_results(self):
+        """Reorganize the results to analyze them.
+
+        TODO must match behavior of get_hypotheses_outputs in parent class
+        """
+        n_sub = len(self.subject_list)
+        tmp = [
+            ("equalIndifference", 1),
+            ("equalRange", 1),
+            ("equalIndifference", 1),
+            ("equalRange", 1),
+            ("equalIndifference", 2),
+            ("equalRange", 2),
+            ("equalIndifference", 2),
+            ("equalRange", 2),
+            ("groupComp", 2),
+        ]
+        h = [
+            join(
+                self.directories.results_dir,
+                self.directories.output_dir,
+                f"l3_analysis_{grp}_nsub_{n_sub}",
+                f"_contrast_id_{contrast}",
+            )
+            for grp, contrast in tmp
         ]
 
-        l3_analysis.connect(
-            [
-                (
-                    subgroups_contrasts,
-                    merge_copes_3rdlevel,
-                    [("copes_global", "in_files")],
-                ),
-                (
-                    subgroups_contrasts,
-                    merge_varcopes_3rdlevel,
-                    [("varcopes_global", "in_files")],
-                ),
-            ]
-        )
-
-    l3_analysis.connect(
-        [
-            (merge_copes_3rdlevel, flame_3rdlevel, [("merged_file", "cope_file")]),
-            (
-                merge_varcopes_3rdlevel,
-                flame_3rdlevel,
-                [("merged_file", "var_cope_file")],
-            ),
-            (
-                specifymodel_3rdlevel,
-                flame_3rdlevel,
-                [
-                    ("design_mat", "design_file"),
-                    ("design_con", "t_con_file"),
-                    ("design_grp", "cov_split_file"),
-                ],
-            ),
-            (flame_3rdlevel, cluster, [("zstats", "in_file"), ("copes", "cope_file")]),
-            (flame_3rdlevel, smoothest, [("zstats", "zstat_file")]),
-            (smoothest, cluster, [("dlh", "dlh"), ("volume", "volume")]),
-            (
-                flame_3rdlevel,
-                datasink_3rdlevel,
-                [
-                    ("zstats", f"l3_analysis_{method}_nsub_{n_sub}.@zstats"),
-                    ("tstats", f"l3_analysis_{method}_nsub_{n_sub}.@tstats"),
-                ],
-            ),
-            (
-                cluster,
-                datasink_3rdlevel,
-                [
-                    ("threshold_file", f"l3_analysis_{method}_nsub_{n_sub}.@thresh"),
-                    ("pval_file", f"l3_analysis_{method}_nsub_{n_sub}.@pval"),
-                ],
-            ),
+        repro_unthresh = [
+            join(filename, "zstat1.nii.gz")
+            if i in [4, 5]
+            else join(filename, "zstat1.nii.gz")
+            for i, filename in enumerate(h)
         ]
-    )
 
-    return l3_analysis
+        repro_thresh = [
+            join(filename, "_cluster0", "zstat1_threshold.nii.gz")
+            if i in [4, 5]
+            else join(filename, "_cluster0", "zstat1_threshold.nii.gz")
+            for i, filename in enumerate(h)
+        ]
 
+        if not os.path.isdir(join(self.directories.results_dir, "NARPS-reproduction")):
+            os.mkdir(join(self.directories.results_dir, "NARPS-reproduction"))
 
-def reorganize_results(result_dir, output_dir, n_sub, team_ID):
-    """
-    Reorganize the results to analyze them.
-
-    Parameters:
-        - result_dir: str, directory where results will be stored
-        - output_dir: str, name of the sub-directory for final results
-        - n_sub: float, number of subject used for the analysis
-        - team_ID: str, ID of the team to reorganize results
-
-    """
-    h1 = opj(
-        result_dir,
-        output_dir,
-        f"l3_analysis_equalIndifference_nsub_{n_sub}",
-        "_contrast_id_1",
-    )
-    h2 = opj(
-        result_dir, output_dir, f"l3_analysis_equalRange_nsub_{n_sub}", "_contrast_id_1"
-    )
-    h3 = opj(
-        result_dir,
-        output_dir,
-        f"l3_analysis_equalIndifference_nsub_{n_sub}",
-        "_contrast_id_1",
-    )
-    h4 = opj(
-        result_dir, output_dir, f"l3_analysis_equalRange_nsub_{n_sub}", "_contrast_id_1"
-    )
-    h5 = opj(
-        result_dir,
-        output_dir,
-        f"l3_analysis_equalIndifference_nsub_{n_sub}",
-        "_contrast_id_2",
-    )
-    h6 = opj(
-        result_dir, output_dir, f"l3_analysis_equalRange_nsub_{n_sub}", "_contrast_id_2"
-    )
-    h7 = opj(
-        result_dir,
-        output_dir,
-        f"l3_analysis_equalIndifference_nsub_{n_sub}",
-        "_contrast_id_2",
-    )
-    h8 = opj(
-        result_dir, output_dir, f"l3_analysis_equalRange_nsub_{n_sub}", "_contrast_id_2"
-    )
-    h9 = opj(
-        result_dir, output_dir, f"l3_analysis_groupComp_nsub_{n_sub}", "_contrast_id_2"
-    )
-
-    h = [h1, h2, h3, h4, h5, h6, h7, h8, h9]
-
-    repro_unthresh = [
-        opj(filename, "zstat1.nii.gz")
-        if i in [4, 5]
-        else opj(filename, "zstat1.nii.gz")
-        for i, filename in enumerate(h)
-    ]
-
-    repro_thresh = [
-        opj(filename, "_cluster0", "zstat1_threshold.nii.gz")
-        if i in [4, 5]
-        else opj(filename, "_cluster0", "zstat1_threshold.nii.gz")
-        for i, filename in enumerate(h)
-    ]
-
-    if not os.path.isdir(opj(result_dir, "NARPS-reproduction")):
-        os.mkdir(opj(result_dir, "NARPS-reproduction"))
-
-    for i, filename in enumerate(repro_unthresh):
-        f_in = filename
-        f_out = opj(
-            result_dir,
-            "NARPS-reproduction",
-            f"team_{team_ID}_nsub_{n_sub}_hypo{i+1}_unthresholded.nii.gz",
-        )
-        shutil.copyfile(f_in, f_out)
-
-    for i, filename in enumerate(repro_thresh):
-        f_in = filename
-        img = nib.load(filename)
-        original_affine = img.affine.copy()
-        spm = nib.load(repro_unthresh[i])
-        new_img = img.get_fdata() > 0.95
-        new_img = new_img.astype(float) * spm.get_fdata()
-        new_spm = nib.Nifti1Image(new_img, original_affine)
-        nib.save(
-            new_spm,
-            opj(
-                result_dir,
+        for i, filename in enumerate(repro_unthresh):
+            f_in = filename
+            f_out = join(
+                self.directories.results_dir,
                 "NARPS-reproduction",
-                f"team_{team_ID}_nsub_{n_sub}_hypo{i+1}_thresholded.nii.gz",
-            ),
-        )
-        # f_out = opj(result_dir, "final_results", f"team_{team_ID}_nsub_{n_sub}_hypo{i+1}_thresholded.nii.gz")
-        # shutil.copyfile(f_in, f_out)
+                f"team_{self.team_id}_nsub_{n_sub}_hypo{i+1}_unthresholded.nii.gz",
+            )
+            shutil.copyfile(f_in, f_out)
 
-    print(f"Results files of team {team_ID} reorganized.")
+        for i, filename in enumerate(repro_thresh):
+            f_in = filename
+            img = nb.load(filename)
+            original_affine = img.affine.copy()
+            spm = nb.load(repro_unthresh[i])
+            new_img = img.get_fdata() > 0.95
+            new_img = new_img.astype(float) * spm.get_fdata()
+            new_spm = nb.Nifti1Image(new_img, original_affine)
+            nb.save(
+                new_spm,
+                join(
+                    self.directories.results_dir,
+                    "NARPS-reproduction",
+                    f"team_{self.team_id}_nsub_{n_sub}_hypo{i+1}_thresholded.nii.gz",
+                ),
+            )
+            # f_out = join(result_dir, "final_results", f"team_{team_ID}_nsub_{n_sub}_hypo{i+1}_thresholded.nii.gz")
+            # shutil.copyfile(f_in, f_out)
+
+        print(f"Results files of team {self.team_id} reorganized.")
+
+        return h
